Fix issue with option parameter parsing
Option parameters with a dash (-) were being considered as options due
to an error with the regular expression used to parse options.
The regex did not specify that the whole word was to be considered, and
so it would do a substring search, and hence identify any parameter
with a dash as a new option.