Enhance type conversion support for args
Arg values could be type cast by specifying a converter delegate to the
TypeAs method.

If a delegate was not specified, the framework would try to find a type
converter for the type.

With this change, if a type converter was not found, the framework will
additionally look for a constructor on the type that can accept a
single string parameter.